requirements.txt
azure-ai-evaluation
python-dotenv
